fix balance type

+++ btcjson/getbalance.go
@@ -1,7 +1,7 @@
 package btcjson
 
 type GetBalanceResult struct {
-	Balance int64
+	Balance float64
 }
 
 type GetBalanceCommand struct {
